Todo와 관련 도메인을 저장하는 로직이 들어나게 이름을 수정하라
