+++ app-server/src/main/java/com/growth/task/todo/application/TodoAddService.java
@@ -1,36 +1,32 @@
 package com.growth.task.todo.application;
 
-import com.growth.task.pomodoro.domain.Pomodoros;
-import com.growth.task.pomodoro.dto.response.PomodoroAddResponse;
-import com.growth.task.pomodoro.service.PomodoroAddService;
+import com.growth.task.task.domain.Tasks;
+import com.growth.task.task.repository.TasksRepository;
 import com.growth.task.todo.domain.Todos;
-import com.growth.task.todo.dto.composite.TodoAndPomodoroAddRequest;
-import com.growth.task.todo.dto.composite.TodoAndPomodoroAddResponse;
-import com.growth.task.todo.dto.response.TodoAddResponse;
+import com.growth.task.todo.dto.request.TodoAddRequest;
+import com.growth.task.todo.exception.TaskNotFoundException;
+import com.growth.task.todo.repository.TodosRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 public class TodoAddService {
+    private TodosRepository todosRepository;
 
-    private final TodoService todoService;
-    private final PomodoroAddService pomodoroAddService;
+    private TasksRepository tasksRepository;
 
-    public TodoAddService(
-            TodoService todoService,
-            PomodoroAddService pomodoroAddService
-    ) {
-        this.todoService = todoService;
-        this.pomodoroAddService = pomodoroAddService;
+    public TodoAddService(TodosRepository todosRepository, TasksRepository tasksRepository) {
+        this.todosRepository = todosRepository;
+        this.tasksRepository = tasksRepository;
     }
 
     @Transactional
-    public TodoAndPomodoroAddResponse save(TodoAndPomodoroAddRequest todoAndPomodoroAddRequest) {
-        Todos todos = todoService.save(todoAndPomodoroAddRequest.getTodoAddRequest());
-        Pomodoros pomodoros = pomodoroAddService.save(todoAndPomodoroAddRequest.getPomodoroAddRequest(), todos);
+    public Todos save(TodoAddRequest todoAddRequest) {
+        Long taskId = todoAddRequest.getTaskId();
+        Tasks tasks = tasksRepository.findById(taskId)
+                .orElseThrow(() -> new TaskNotFoundException(taskId));
 
-        TodoAddResponse todoAddResponse = new TodoAddResponse(todos);
-        PomodoroAddResponse pomodoroAddResponse = new PomodoroAddResponse(pomodoros);
-        return new TodoAndPomodoroAddResponse(todoAddResponse, pomodoroAddResponse);
+        Todos todos = todoAddRequest.toEntity(tasks);
+        return todosRepository.save(todos);
     }
 }
