+++ app-server/src/test/java/com/growth/task/task/service/TaskAddServiceTest.java
@@ -36,7 +36,7 @@ class TaskAddServiceTest {
     }
 
     @Nested
-    @DisplayName("save")
+    @DisplayName("saveWithRelated")
     class Describe_save {
         @Nested
         @DisplayName("Task 생성 정보가 주어지면")
