+++ app-server/src/test/java/com/growth/task/todo/application/TodoAddServiceTest.java
@@ -1,14 +1,11 @@
 package com.growth.task.todo.application;
 
-import com.growth.task.pomodoro.domain.Pomodoros;
-import com.growth.task.pomodoro.dto.request.PomodoroAddRequest;
-import com.growth.task.pomodoro.service.PomodoroAddService;
 import com.growth.task.task.domain.Tasks;
+import com.growth.task.task.repository.TasksRepository;
 import com.growth.task.todo.domain.Todos;
-import com.growth.task.todo.dto.composite.TodoAndPomodoroAddRequest;
-import com.growth.task.todo.dto.composite.TodoAndPomodoroAddResponse;
 import com.growth.task.todo.dto.request.TodoAddRequest;
-import com.growth.task.todo.enums.Status;
+import com.growth.task.todo.exception.TaskNotFoundException;
+import com.growth.task.todo.repository.TodosRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -17,94 +14,83 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.lenient;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.times;
+import java.util.Optional;
+
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class TodoAddServiceTest {
 
     @Mock
-    private TodoService todoService;
+    private TodosRepository todosRepository;
     @Mock
-    private PomodoroAddService pomodoroAddService;
+    private TasksRepository tasksRepository;
 
     private TodoAddService todoAddService;
 
-    private final Long TASK_ID = 1L;
-    private final Long TODO_ID1 = 1L;
-    private final String WHAT_TO_DO1 = "디자인패턴의 아름다움 스터디";
-    private final int POMODORO_PERFORM_COUNT1 = 0;
-    private final int POMODORO_PLAN_COUNT1 = 1;
+    private final Long TASK_ID1 = 1L;
+    private final Long TASK_ID2 = 2L;
+    private final String WHAT_TO_DO = "디자인패턴의 아름다움 스터디";
 
     @BeforeEach
     void setUp() {
-        todoService = mock(TodoService.class);
-        pomodoroAddService = mock(PomodoroAddService.class);
-        todoAddService = new TodoAddService(todoService, pomodoroAddService);
+        todoAddService = new TodoAddService(todosRepository, tasksRepository);
+
     }
 
     @Nested
-    @DisplayName("signUp 메서드는")
+    @DisplayName("TodoService 의 signUp 메서드는")
     class Describe_save {
 
         @Nested
-        @DisplayName("Todo 및 Pomodoro 정보가 주어지면")
-        class Context_whenTaskAndPomodoroExists {
-            private final Tasks tasks = Tasks.builder()
-                    .taskId(TASK_ID)
-                    .build();
-            private final Todos todos = Todos.builder()
-                    .todoId(TODO_ID1)
-                    .task(tasks)
-                    .todo(WHAT_TO_DO1)
-                    .status(Status.READY)
-                    .build();
-            private final Pomodoros pomodoro = Pomodoros.builder()
-                    .todo(todos)
-                    .performCount(POMODORO_PERFORM_COUNT1)
-                    .planCount(POMODORO_PLAN_COUNT1)
-                    .build();
+        @DisplayName("taskId 로 task 가 확인되는 경우")
+        class Context_whenTaskIdExists {
             private final TodoAddRequest todoAddRequest = TodoAddRequest.builder()
-                    .taskId(tasks.getTaskId())
-                    .todo(WHAT_TO_DO1)
+                    .taskId(TASK_ID1)
+                    .todo(WHAT_TO_DO)
                     .build();
-            private final PomodoroAddRequest pomodoroAddRequest = PomodoroAddRequest.builder()
-                    .performCount(POMODORO_PERFORM_COUNT1)
-                    .planCount(POMODORO_PLAN_COUNT1)
-                    .build();
-            private final TodoAndPomodoroAddRequest todoAndPomodoroAddRequest = TodoAndPomodoroAddRequest.builder()
-                    .todoAddRequest(todoAddRequest)
-                    .pomodoroAddRequest(pomodoroAddRequest)
+
+            private final Tasks tasks = Tasks.builder()
+                    .taskId(TASK_ID1)
                     .build();
 
             @BeforeEach
             void setUp() {
-                lenient().when(todoService.save(any(TodoAddRequest.class))).thenReturn(todos);
-                lenient().when(pomodoroAddService.save(any(PomodoroAddRequest.class), any(Todos.class))).thenReturn(pomodoro);
+                given(tasksRepository.findById(TASK_ID1)).willReturn(Optional.of(tasks));
             }
 
             @Test
-            @DisplayName("정상적으로 저장하고 CompositeAddResponse 를 반환한다.")
-            void It_shouldSaveAndReturnCompositeResponse() {
-                TodoAndPomodoroAddResponse response = todoAddService.save(todoAndPomodoroAddRequest);
+            @DisplayName("todoAddRequest 가 저장된다.")
+            void It_saveTheTodo() {
+                todoAddService.save(todoAddRequest);
 
-                assertAll(
-                        () -> assertThat(response.getTodoId()).isEqualTo(TODO_ID1),
-                        () -> assertThat(response.getTaskId()).isEqualTo(TASK_ID),
-                        () -> assertThat(response.getTodo()).isEqualTo(WHAT_TO_DO1),
-                        () -> assertThat(response.getStatus()).isEqualTo(Status.READY.toString()),
-                        () -> assertThat(response.getPerformCount()).isEqualTo(POMODORO_PERFORM_COUNT1),
-                        () -> assertThat(response.getPlanCount()).isEqualTo(POMODORO_PLAN_COUNT1)
-                );
+                verify(todosRepository).save(any(Todos.class));
+            }
+        }
 
-                verify(todoService, times(1)).save(any(TodoAddRequest.class));
-                verify(pomodoroAddService, times(1)).save(any(PomodoroAddRequest.class), any(Todos.class));
+        @Nested
+        @DisplayName("taskId 로 Tasks 가 확인되지 않는 경우")
+        class Context_whenTaskIdDoesNotExist {
+            private final TodoAddRequest todoAddRequest = TodoAddRequest.builder()
+                    .taskId(TASK_ID2)
+                    .todo(WHAT_TO_DO)
+                    .build();
+
+            @BeforeEach
+            void setUp() {
+                when(tasksRepository.findById(TASK_ID2)).thenReturn(Optional.empty());
+            }
+
+            @Test
+            @DisplayName("TaskNotFoundException 오류를 던진다.")
+            void It_throws_TaskNotFoundException() {
+                assertThrows(TaskNotFoundException.class, () -> todoAddService.save(todoAddRequest));
             }
         }
     }
+
 }
